+++ client/kpm-loggedin.js.handlebars
