+++ server/server.js
@@ -47,12 +47,12 @@ function compileTemplate(name) {
   );
 }
 
-const template = compileTemplate("index.handlebars");
-const loggedInTemplate = compileTemplate("kpm-loggedin.js.handlebars");
-const loggedOutTemplate = compileTemplate("kpm-loggedout.js.handlebars");
+const infoPageTemplate = compileTemplate("info-page.handlebars");
+const loggedInTemplate = compileTemplate("../client/kpm-loggedin.js.handlebars");
+const loggedOutTemplate = compileTemplate("../client/kpm-loggedout.js.handlebars");
 
 // TODO: Sass?
-const menuCssData = fs.readFileSync(path.resolve(__dirname, "menu.css"));
+const menuCssData = fs.readFileSync(path.resolve(__dirname, "../client/menu.css"));
 const menuCssName = `menu-${hash(menuCssData)}.css`;
 
 function hash(data) {
@@ -142,7 +142,7 @@ app.get("/kpm", async (req, res) => {
   const search = await fetchBlock("search");
 
   res.send(
-    template({
+    infoPageTemplate({
       footer,
       megaMenu,
       search,
